+++ java/client/src/main/java/glide/managers/ConnectionManager.java
@@ -467,6 +467,7 @@ public class ConnectionManager {
                             cacheBuilder.setCacheId(clientSideCache.getCacheId());
                             cacheBuilder.setMaxCacheKb(clientSideCache.getMaxCacheKb());
                             cacheBuilder.setEnableMetrics(clientSideCache.isEnableMetrics());
+                            cacheBuilder.setServerAssisted(clientSideCache.isServerAssisted());
 
                             // Set TTL (0 = no expiration)
                             cacheBuilder.setEntryTtlMs(clientSideCache.getEntryTtlMs());